feat: make azure experimental the default
