+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/Field.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import { EditorField } from '@grafana/experimental';
-import { config } from '@grafana/runtime';
 import { InlineField } from '@grafana/ui';
 import { Props as InlineFieldProps } from '@grafana/ui/src/components/Forms/InlineField';
 
@@ -16,7 +15,7 @@ const DEFAULT_LABEL_WIDTH = 18;
 export const Field = (props: Props) => {
   const { labelWidth, inlineField, ...remainingProps } = props;
 
-  if (config.featureToggles.azureMonitorExperimentalUI && !inlineField) {
+  if (!inlineField) {
     return <EditorField width={labelWidth || DEFAULT_LABEL_WIDTH} {...remainingProps} />;
   } else {
     return <InlineField labelWidth={labelWidth || DEFAULT_LABEL_WIDTH} {...remainingProps} />;
