+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/DimensionFields.tsx
@@ -1,12 +1,13 @@
 import React, { useEffect, useMemo, useState } from 'react';
 
-import { SelectableValue, DataFrame, PanelData } from '@grafana/data';
-import { Button, Select, HorizontalGroup, VerticalGroup, MultiSelect } from '@grafana/ui';
+import { SelectableValue, DataFrame, PanelData, Labels } from '@grafana/data';
+import { AccessoryButton, EditorList } from '@grafana/experimental';
+import { Select, HorizontalGroup, MultiSelect } from '@grafana/ui';
 
 import { AzureMetricDimension, AzureMonitorOption, AzureMonitorQuery, AzureQueryEditorFieldProps } from '../../types';
 import { Field } from '../Field';
 
-import { appendDimensionFilter, removeDimensionFilter, setDimensionFilterValue } from './setQueryValue';
+import { setDimensionFilters } from './setQueryValue';
 
 interface DimensionFieldsProps extends AzureQueryEditorFieldProps {
   dimensionOptions: AzureMonitorOption[];
@@ -28,16 +29,14 @@ const useDimensionLabels = (data: PanelData | undefined, query: AzureMonitorQuer
       const labels = fields
         .map((fields) => fields.labels)
         .flat()
-        .filter((item) => item!);
+        .filter((item): item is Labels => item !== null && item !== undefined);
       for (const label of labels) {
         // Labels only exist for series that have a dimension selected
-        if (label) {
-          for (const [dimension, value] of Object.entries(label)) {
-            if (labelsObj[dimension]) {
-              labelsObj[dimension].add(value);
-            } else {
-              labelsObj[dimension] = new Set([value]);
-            }
+        for (const [dimension, value] of Object.entries(label)) {
+          if (labelsObj[dimension]) {
+            labelsObj[dimension].add(value);
+          } else {
+            labelsObj[dimension] = new Set([value]);
           }
         }
       }
@@ -87,24 +86,14 @@ const DimensionFields: React.FC<DimensionFieldsProps> = ({ data, query, dimensio
     return t;
   }, [dimensionFilters, dimensionOptions]);
 
-  const addFilter = () => {
-    onQueryChange(appendDimensionFilter(query));
-  };
-
-  const removeFilter = (index: number) => {
-    onQueryChange(removeDimensionFilter(query, index));
-  };
-
   const onFieldChange = <Key extends keyof AzureMetricDimension>(
-    filterIndex: number,
     fieldName: Key,
-    value: AzureMetricDimension[Key]
+    item: Partial<AzureMetricDimension>,
+    value: AzureMetricDimension[Key],
+    onChange: (item: Partial<AzureMetricDimension>) => void
   ) => {
-    onQueryChange(setDimensionFilterValue(query, filterIndex, fieldName, value));
-  };
-
-  const onFilterInputChange = (index: number, v: SelectableValue<string> | null) => {
-    onFieldChange(index, 'filters', [v?.value ?? '']);
+    item[fieldName] = value;
+    onChange(item);
   };
 
   const getValidDimensionOptions = (selectedDimension: string) => {
@@ -133,7 +122,6 @@ const DimensionFields: React.FC<DimensionFieldsProps> = ({ data, query, dimensio
     }
     return labelOptions;
   };
-
   const getValidOperators = (selectedOperator: string) => {
     if (dimensionOperators.find((operator: SelectableValue) => operator.value === selectedOperator)) {
       return dimensionOperators;
@@ -141,70 +129,76 @@ const DimensionFields: React.FC<DimensionFieldsProps> = ({ data, query, dimensio
     return [...dimensionOperators, ...(selectedOperator ? [{ label: selectedOperator, value: selectedOperator }] : [])];
   };
 
-  const onMultiSelectFilterChange = (index: number, v: Array<SelectableValue<string>>) => {
-    onFieldChange(
-      index,
-      'filters',
-      v.map((item) => item.value || '')
+  const changedFunc = (changed: Array<Partial<AzureMetricDimension>>) => {
+    const properData: AzureMetricDimension[] = changed.map((x) => {
+      return {
+        dimension: x.dimension ?? '',
+        operator: x.operator ?? 'eq',
+        filters: x.filters ?? [],
+      };
+    });
+    onQueryChange(setDimensionFilters(query, properData));
+  };
+
+  const renderFilters = (
+    item: Partial<AzureMetricDimension>,
+    onChange: (item: Partial<AzureMetricDimension>) => void,
+    onDelete: () => void
+  ) => {
+    return (
+      <HorizontalGroup spacing="none">
+        <Select
+          menuShouldPortal
+          placeholder="Field"
+          value={item.dimension}
+          options={getValidDimensionOptions(item.dimension || '')}
+          onChange={(e) => onFieldChange('dimension', item, e.value ?? '', onChange)}
+        />
+        <Select
+          menuShouldPortal
+          placeholder="Operation"
+          value={item.operator}
+          options={getValidOperators(item.operator || 'eq')}
+          onChange={(e) => onFieldChange('operator', item, e.value ?? '', onChange)}
+          allowCustomValue
+        />
+        {item.operator === 'eq' || item.operator === 'ne' ? (
+          <MultiSelect
+            menuShouldPortal
+            placeholder="Select value(s)"
+            value={item.filters}
+            options={getValidMultiSelectOptions(item.filters, item.dimension ?? '')}
+            onChange={(e) =>
+              onFieldChange(
+                'filters',
+                item,
+                e.map((x) => x.value ?? ''),
+                onChange
+              )
+            }
+            aria-label={'dimension-labels-select'}
+            allowCustomValue
+          />
+        ) : (
+          // The API does not currently allow for multiple "starts with" clauses to be used.
+          <Select
+            menuShouldPortal
+            placeholder="Select value"
+            value={item.filters ? item.filters[0] : ''}
+            allowCustomValue
+            options={getValidFilterOptions(item.filters ? item.filters[0] : '', item.dimension ?? '')}
+            onChange={(e) => onFieldChange('filters', item, [e?.value ?? ''], onChange)}
+            isClearable
+          />
+        )}
+        <AccessoryButton aria-label="Remove" icon="times" variant="secondary" onClick={onDelete} type="button" />
+      </HorizontalGroup>
     );
   };
 
   return (
-    <Field label="Dimension">
-      <VerticalGroup spacing="xs">
-        {dimensionFilters.map((filter, index) => (
-          <HorizontalGroup key={index} spacing="xs">
-            <Select
-              placeholder="Field"
-              value={filter.dimension}
-              options={getValidDimensionOptions(filter.dimension)}
-              onChange={(v) => onFieldChange(index, 'dimension', v.value ?? '')}
-              width={38}
-            />
-            <Select
-              menuShouldPortal
-              placeholder="Operation"
-              value={filter.operator}
-              options={getValidOperators(filter.operator)}
-              onChange={(v) => onFieldChange(index, 'operator', v.value ?? '')}
-              allowCustomValue
-            />
-            {filter.operator === 'eq' || filter.operator === 'ne' ? (
-              <MultiSelect
-                menuShouldPortal
-                placeholder="Select value(s)"
-                value={filter.filters}
-                options={getValidMultiSelectOptions(filter.filters, filter.dimension)}
-                onChange={(v) => onMultiSelectFilterChange(index, v)}
-                aria-label={'dimension-labels-select'}
-                allowCustomValue
-              />
-            ) : (
-              // The API does not currently allow for multiple "starts with" clauses to be used.
-              <Select
-                menuShouldPortal
-                placeholder="Select value"
-                value={filter.filters ? filter.filters[0] : ''}
-                allowCustomValue
-                options={getValidFilterOptions(filter.filters ? filter.filters[0] : '', filter.dimension)}
-                onChange={(v) => onFilterInputChange(index, v)}
-                isClearable
-              />
-            )}
-            <Button
-              variant="secondary"
-              size="md"
-              icon="trash-alt"
-              aria-label="Remove"
-              onClick={() => removeFilter(index)}
-            ></Button>
-          </HorizontalGroup>
-        ))}
-
-        <Button variant="secondary" size="md" onClick={addFilter}>
-          Add new dimension
-        </Button>
-      </VerticalGroup>
+    <Field label="Dimensions">
+      <EditorList items={dimensionFilters} onChange={changedFunc} renderItem={renderFilters} />
     </Field>
   );
 };
