+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/MetricsQueryEditor.tsx
@@ -3,8 +3,6 @@ import React from 'react';
 
 import { PanelData } from '@grafana/data/src/types';
 import { EditorRows, EditorRow, EditorFieldGroup } from '@grafana/experimental';
-import { config } from '@grafana/runtime';
-import { InlineFieldRow, useStyles2 } from '@grafana/ui';
 
 import type Datasource from '../../datasource';
 import type { AzureMonitorQuery, AzureMonitorOption, AzureMonitorErrorish } from '../../types';
@@ -16,7 +14,6 @@ import DimensionFields from './DimensionFields';
 import LegendFormatField from './LegendFormatField';
 import MetricNameField from './MetricNameField';
 import MetricNamespaceField from './MetricNamespaceField';
-import NewDimensionFields from './NewDimensionFields';
 import TimeGrainField from './TimeGrainField';
 import TopField from './TopField';
 import { useMetricNames, useMetricNamespaces, useMetricMetadata } from './dataHooks';
@@ -39,181 +36,98 @@ const MetricsQueryEditor: React.FC<MetricsQueryEditorProps> = ({
   onChange,
   setError,
 }) => {
-  const styles = useStyles2(getStyles);
-
   const metricsMetadata = useMetricMetadata(query, datasource, onChange);
   const metricNamespaces = useMetricNamespaces(query, datasource, onChange, setError);
   const metricNames = useMetricNames(query, datasource, onChange, setError);
-  if (config.featureToggles.azureMonitorExperimentalUI) {
-    return (
-      <span data-testid="azure-monitor-metrics-query-editor-with-experimental-ui">
-        <EditorRows>
-          <EditorRow>
-            <EditorFieldGroup>
-              <ResourceField
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-                selectableEntryTypes={[ResourceRowType.Resource]}
-                setResource={setResource}
-                resourceUri={query.azureMonitor?.resourceUri}
-                queryType={'metrics'}
-              />
-              <MetricNamespaceField
-                metricNamespaces={metricNamespaces}
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-              />
-              <MetricNameField
-                metricNames={metricNames}
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-              />
-            </EditorFieldGroup>
-          </EditorRow>
-          <EditorRow>
-            <EditorFieldGroup>
-              <AggregationField
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-                aggregationOptions={metricsMetadata?.aggOptions ?? []}
-                isLoading={metricsMetadata.isLoading}
-              />
-              <TimeGrainField
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-                timeGrainOptions={metricsMetadata?.timeGrains ?? []}
-              />
-            </EditorFieldGroup>
-          </EditorRow>
-          <EditorRow>
-            <EditorFieldGroup>
-              <NewDimensionFields
-                data={data}
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-                dimensionOptions={metricsMetadata?.dimensions ?? []}
-              />
-            </EditorFieldGroup>
-          </EditorRow>
-          <EditorRow>
-            <EditorFieldGroup>
-              <TopField
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-              />
-              <LegendFormatField
-                query={query}
-                datasource={datasource}
-                variableOptionGroup={variableOptionGroup}
-                onQueryChange={onChange}
-                setError={setError}
-              />
-            </EditorFieldGroup>
-          </EditorRow>
-        </EditorRows>
-      </span>
-    );
-  } else {
-    return (
-      <div data-testid="azure-monitor-metrics-query-editor-with-resource-picker">
-        <InlineFieldRow className={styles.row}>
-          <ResourceField
-            query={query}
-            datasource={datasource}
-            variableOptionGroup={variableOptionGroup}
-            onQueryChange={onChange}
-            setError={setError}
-            selectableEntryTypes={[ResourceRowType.Resource]}
-            setResource={setResource}
-            resourceUri={query.azureMonitor?.resourceUri}
-            queryType="metrics"
-          />
-        </InlineFieldRow>
-
-        <InlineFieldRow className={styles.row}>
-          <MetricNamespaceField
-            metricNamespaces={metricNamespaces}
-            query={query}
-            datasource={datasource}
-            variableOptionGroup={variableOptionGroup}
-            onQueryChange={onChange}
-            setError={setError}
-          />
-          <MetricNameField
-            metricNames={metricNames}
-            query={query}
-            datasource={datasource}
-            variableOptionGroup={variableOptionGroup}
-            onQueryChange={onChange}
-            setError={setError}
-          />
-        </InlineFieldRow>
-        <InlineFieldRow className={styles.row}>
-          <AggregationField
-            query={query}
-            datasource={datasource}
-            variableOptionGroup={variableOptionGroup}
-            onQueryChange={onChange}
-            setError={setError}
-            aggregationOptions={metricsMetadata?.aggOptions ?? []}
-            isLoading={metricsMetadata.isLoading}
-          />
-          <TimeGrainField
-            query={query}
-            datasource={datasource}
-            variableOptionGroup={variableOptionGroup}
-            onQueryChange={onChange}
-            setError={setError}
-            timeGrainOptions={metricsMetadata?.timeGrains ?? []}
-          />
-        </InlineFieldRow>
-        <DimensionFields
-          data={data}
-          query={query}
-          datasource={datasource}
-          variableOptionGroup={variableOptionGroup}
-          onQueryChange={onChange}
-          setError={setError}
-          dimensionOptions={metricsMetadata?.dimensions ?? []}
-        />
-        <TopField
-          query={query}
-          datasource={datasource}
-          variableOptionGroup={variableOptionGroup}
-          onQueryChange={onChange}
-          setError={setError}
-        />
-        <LegendFormatField
-          query={query}
-          datasource={datasource}
-          variableOptionGroup={variableOptionGroup}
-          onQueryChange={onChange}
-          setError={setError}
-        />
-      </div>
-    );
-  }
+  return (
+    <span data-testid="azure-monitor-metrics-query-editor-with-experimental-ui">
+      <EditorRows>
+        <EditorRow>
+          <EditorFieldGroup>
+            <ResourceField
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+              selectableEntryTypes={[ResourceRowType.Resource]}
+              setResource={setResource}
+              resourceUri={query.azureMonitor?.resourceUri}
+              queryType={'metrics'}
+            />
+            <MetricNamespaceField
+              metricNamespaces={metricNamespaces}
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+            />
+            <MetricNameField
+              metricNames={metricNames}
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+            />
+          </EditorFieldGroup>
+        </EditorRow>
+        <EditorRow>
+          <EditorFieldGroup>
+            <AggregationField
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+              aggregationOptions={metricsMetadata?.aggOptions ?? []}
+              isLoading={metricsMetadata.isLoading}
+            />
+            <TimeGrainField
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+              timeGrainOptions={metricsMetadata?.timeGrains ?? []}
+            />
+          </EditorFieldGroup>
+        </EditorRow>
+        <EditorRow>
+          <EditorFieldGroup>
+            <DimensionFields
+              data={data}
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+              dimensionOptions={metricsMetadata?.dimensions ?? []}
+            />
+          </EditorFieldGroup>
+        </EditorRow>
+        <EditorRow>
+          <EditorFieldGroup>
+            <TopField
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+            />
+            <LegendFormatField
+              query={query}
+              datasource={datasource}
+              variableOptionGroup={variableOptionGroup}
+              onQueryChange={onChange}
+              setError={setError}
+            />
+          </EditorFieldGroup>
+        </EditorRow>
+      </EditorRows>
+    </span>
+  );
 };
 
 const getStyles = () => ({
